First Major Commit
This is for the very early prototype/feasibility study.

The platformio.ini is configured such that com[4] is connected to the ESP8266.

Make sure the WifiManager library is installed.

Steps to use:
Run the python server (TCP_Server.py)

change all network settings in both the python server and esp client so that it suits your network (its hard coded for mine right now)

you will need to connect to you wifi first before you can see WiFi Manager (see commented out code in main)

Turn on the ESP8266

If successful, the TCP_Server should print out lines when recieved.

There is alot of redundant delay and waiting in the ESP8266 file so the sending of data is about 1 second apart.

TODO:
-Use serial to get MPU6050 values from FLORA
-Send these values to the TCP server
-Basic localisation using trilateration (we can start in offline mode using a log, live tracking will be simmilar and can be done later)